MiniProjectwork.md
The website is a virtual pet clinic and retail store that sells pet products. There is the possibility to book an appointment for your pets by giving us their preferred time and date, as well as to book online or remotely. It is possible to purchase all the necessities for your pet, such as food, medicine, toys, etc.